rs: Fix build with ISO C++ 11 compilers
ISO C++ 11 no longer allows implicit narrowing during initialization

+++ cpu_ref/rsCpuIntrinsic3DLUT.cpp
@@ -70,10 +70,10 @@ void RsdCpuScriptIntrinsic3DLUT::kernel(const RsForEachStubParamStruct *p,
     const uchar *bp = (const uchar *)cp->mLUT->mHal.drvState.lod[0].mallocPtr;
 
     int4 dims = {
-        cp->mLUT->mHal.drvState.lod[0].dimX - 1,
-        cp->mLUT->mHal.drvState.lod[0].dimY - 1,
-        cp->mLUT->mHal.drvState.lod[0].dimZ - 1,
-        -1
+        static_cast<int>(cp->mLUT->mHal.drvState.lod[0].dimX - 1),
+        static_cast<int>(cp->mLUT->mHal.drvState.lod[0].dimY - 1),
+        static_cast<int>(cp->mLUT->mHal.drvState.lod[0].dimZ - 1),
+        static_cast<int>(-1)
     };
     const float4 m = (float4)(1.f / 255.f) * convert_float4(dims);
     const int4 coordMul = convert_int4(m * (float4)0x8000);
@@ -87,8 +87,8 @@ void RsdCpuScriptIntrinsic3DLUT::kernel(const RsForEachStubParamStruct *p,
         int32_t len = (x2 - x1 - 1) >> 1;
         if(len > 0) {
             const short neon_constants[] = {
-                coordMul.x, coordMul.y, coordMul.z, 0,
-                0, 0, 0, 0xffff,
+                static_cast<short>(coordMul.x), static_cast<short>(coordMul.y), static_cast<short>(coordMul.z), static_cast<short>(0),
+                0, 0, 0, static_cast<short>(0xffff),
 
             };
